When running the concurrency suite, sequentialize the plotting
The final step of running each concurrency test ends up running inkscape
to convert the SVG files into PNGs. This is a rather large program and
ends up being run in parallel once for each core, which can rapidly
exhaust memory on machines with high core counts. This moves that step
out of the parallel section.